requirements.txt
pandas
numpy
scikit-learn
seaborn
matplotlib
xgboost
lightgbm
joblib
streamlit
python-dotenv
